refactor(bsw): split Cpu0_Main into modules, move cross-core data to LMU
Cpu0_Main.c was 630 lines and carried computation, orchestration and
boot-time debug that belong elsewhere. Worse, everything with a hard
deadline shared a cooperative core with everything that blocks: the
50 Hz estimator sat next to two I2C reads with a 10 ms deadline each
and a DFLASH erase that stalls the core for ~100 ms.

Implements T1-T11 of docs/REFACTORING_PLAN.md. Cpu0_Main.c is now
303 lines of entry, init and task registration.

New modules, one responsibility each:
  BringUp      boot-only register dumps over UART
  Atmosphere   pure: pressure -> altitude, QNH clamp
  SensorTask   baro/mag/GNSS periodic read, convert, publish, diagnose
  Housekeeping off-chain 100 ms: measurements, diagnostics, gpio, load
  NavTask      the flight chain: IMU -> AHRS -> fusion -> publish
  SharedRam    the LMU cross-core block, __at() definitions only
  NavState     CPU1 -> CPU0 snapshot, generation counter

Cross-core data now follows Infineon's documented mechanism instead of
the reader-side-only assumption in CoreStats.h: shared objects live in
the LMU at 0xB00F0000, addressed through the non-cached 0xB alias, one
writer per object, publish = payload -> Ifx__dsync() -> gen++. The
barrier is verified as a real DSYNC instruction in the generated .src,
and the block's placement is verified in the .map, because a missing
barrier fails silently as a rare torn read rather than as a build
error. Ifx__dsync is absent from IfxCpu_IntrinsicsTasking.h and is
wrapped in SharedRam.h; recorded in docs/ILLD_NOTES.md.

g_coreStats is the block's first consumer, deliberately: six writers
already read live over XCP make the loudest possible canary if the
contract is wrong. loadPmil and aliveCounter widen to uint32 because
LMU SRAM is 64-bit wide with no sub-word access; Xcp_Data keeps its
uint16 fields and its offsets, so the A2L is unchanged.

Two latent NaN bugs fixed in passing, both of the same shape as the
one that previously killed a fusion channel - a range check written
as (x < lo || x > hi) admits NaN because every comparison against NaN
is false:
  - Atmosphere.c QNH clamp
  - NavTask dt window

Also carries I1 and I3 of docs/IMU_INTERRUPT.md: P10.7 is configured
input-only with a pulldown and a TTL pad driver before any other pin
setup, because it is a 5 V VEXT pad wired to INT1 whose absolute
maximum is 3.6 V; and the ICM-42688-P INT1 registers are written as
whole bytes, since INT_CONFIG1 and INT_SOURCE0 both reset to 0x10 and
a read-modify-write would preserve INT_ASYNC_RESET=1, which silently
prevents any interrupt from ever firing.

Verified on hardware: |a| = 1.0018 g, diagStatus 0x800 (benign UART
bit only), NavDropped and NavCovResets 0 over 5 min, all six core
alive-counters incrementing in lockstep. 15/15 host tests, MISRA
0 findings, build 0 errors 0 warnings.

Not verified: the AHRS nose-up bench check needs the board tilted by
hand, and NavState_get's torn-read retry is unreachable from a
single-threaded host test. Both are disclosed in place.

+++ src/bsw/Measurements.c
@@ -107,9 +107,17 @@ void measurementsSetSystemLoad(void)
 
     for (i = 0u; i < CORESTATS_NUM_CORES; i++)
     {
+        /* coreLoadPmil/coreAlive stay uint16 in Xcp_Data -- this block is
+         * plain CPU0 DSPR, not the LMU, so the T9 no-sub-word rule (32-bit
+         * fields only in CoreStats_t, SharedRam.h) does not apply here, and
+         * widening the XCP-visible copies would cost 24 bytes this block
+         * cannot spare (docs/REFACTORING_PLAN.md §2.4, T9 acceptance) for no
+         * benefit: loadPmil never exceeds 1000 and aliveCounter already
+         * wrapped at 65536 before this change. Explicit casts, not implicit
+         * narrowing (MISRA 10.3/10.8). */
         g_xcpData.coreExecUs[i]   = g_coreStats[i].execUs;
-        g_xcpData.coreLoadPmil[i] = g_coreStats[i].loadPmil;
-        g_xcpData.coreAlive[i]    = g_coreStats[i].aliveCounter;
+        g_xcpData.coreLoadPmil[i] = (uint16)g_coreStats[i].loadPmil;
+        g_xcpData.coreAlive[i]    = (uint16)g_coreStats[i].aliveCounter;
     }
 
     EthStats_update();
